*fix* fix segmentation fault. We need to specify real*8 for the voxel field, the default seems to be single precision

+++ fortran/mr_protocol.f90
@@ -7,7 +7,7 @@ implicit none
 ! 
 type MRI
      ! voxel_feature
-     real, dimension(:,:), allocatable :: voxel_feature
+     real*8, dimension(:,:), allocatable :: voxel_feature
      integer, dimension(2) :: voxel_feature_dims
 end type
